Implement ScrollableBackground and add images
ScrollableBackground is designed to deal with scrolling backgrounds to
simulate the effect of parallax. As of now, 4 different layers of
parallax are being drawn, however, the game runs extremely slowly.
Played around with the Memory Allocation Tracker, but wasn't able to
find anything. Speeding this up is a number 1 priority.

Also, my pixel art that I drew using Aseprite was added.